Added Asynchronous Success function support
If asyncSuccess is true then the success function will be called with the controller(this) as the first parameter and the close dialog function as the second parameter.

This allows you to conditionally continue. In my case I needed the success to not close the dialog if perhaps the data entered in the dialog was invalid. Also I needed to validate asynchronously.


FYI if this is not something fundoo-solutions doesn't want in createDialog thats fine. I just needed it for my project. :-)